# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FreshCart API
  description: |
    REST API for the FreshCart online grocery platform.
    
    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `/api/auth/login` or `/api/auth/register`.
    Include it as: `Authorization: Bearer <token>`
    
    ## Roles
    - `customer` — can browse, cart, order, wishlist
    - `admin` — full platform management
    - `delivery_partner` — view assigned orders, update status
  version: 1.0.0
  contact:
    name: FreshCart API Support
    email: api@freshcart.example.com
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: https://api.freshcart.example.com/api
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [customer, admin, delivery_partner]
        isActive:
          type: boolean
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      summary: Register with email/password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  example: Alice Johnson
                email:
                  type: string
                  format: email
                  example: alice@example.com
                password:
                  type: string
                  minLength: 8
                  example: Password123!
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          user:
                            $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '409':
          description: Email already registered
  /auth/login:
    post:
      summary: Login with email/password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New token pair issued
        '401':
          description: Invalid or expired refresh token
  /auth/logout:
    post:
      summary: Logout (invalidate refresh token)
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logged out successfully
...
  /cart:
    get:
      summary: Get cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart with items and subtotal
  /cart/items:
    post:
      summary: Add item to cart
      tags: [Cart]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  default: 1
      responses:
        '200':
          description: Item added
  /orders:
    get:
      summary: Get order history
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated order history
    post:
      summary: Place order
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [addressId, deliverySlotId, paymentMethod]
              properties:
                addressId:
                  type: string
                  format: uuid
                deliverySlotId:
                  type: string
                  format: uuid
                paymentMethod:
                  type: string
                  enum: [upi, card, cod]
      responses:
        '201':
          description: Order placed
        '400':
          description: Validation error or out-of-stock items
...